+++ homeassistant/components/device_automation/__init__.py
@@ -1,21 +1,23 @@
 """Helpers for device automations."""
 import asyncio
 import logging
-from typing import Callable, cast
 
 import voluptuous as vol
 
 from homeassistant.components import websocket_api
-from homeassistant.const import CONF_DOMAIN
-from homeassistant.core import split_entity_id, HomeAssistant
-import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_registry import async_entries_for_device
-from homeassistant.helpers.typing import ConfigType
 from homeassistant.loader import async_get_integration, IntegrationNotFound
 
 DOMAIN = "device_automation"
 
 _LOGGER = logging.getLogger(__name__)
+
+
+TYPES = {
+    "trigger": ("device_trigger", "async_get_triggers"),
+    "condition": ("device_condition", "async_get_conditions"),
+    "action": ("device_action", "async_get_actions"),
+}
 
 
 async def async_setup(hass, config):
@@ -32,21 +34,9 @@ async def async_setup(hass, config):
     return True
 
 
-async def async_device_condition_from_config(
-    hass: HomeAssistant, config: ConfigType, config_validation: bool = True
-) -> Callable[..., bool]:
-    """Wrap action method with state based condition."""
-    if config_validation:
-        config = cv.DEVICE_CONDITION_SCHEMA(config)
-    integration = await async_get_integration(hass, config[CONF_DOMAIN])
-    platform = integration.get_platform("device_automation")
-    return cast(
-        Callable[..., bool],
-        platform.async_condition_from_config(config, config_validation),  # type: ignore
-    )
-
-
-async def _async_get_device_automations_from_domain(hass, domain, fname, device_id):
+async def _async_get_device_automations_from_domain(
+    hass, domain, automation_type, device_id
+):
     """List device automations."""
     integration = None
     try:
@@ -55,17 +45,18 @@ async def _async_get_device_automations_from_domain(hass, domain, fname, device_
         _LOGGER.warning("Integration %s not found", domain)
         return None
 
+    platform_name, function_name = TYPES[automation_type]
+
     try:
-        platform = integration.get_platform("device_automation")
+        platform = integration.get_platform(platform_name)
     except ImportError:
         # The domain does not have device automations
         return None
 
-    if hasattr(platform, fname):
-        return await getattr(platform, fname)(hass, device_id)
+    return await getattr(platform, function_name)(hass, device_id)
 
 
-async def _async_get_device_automations(hass, fname, device_id):
+async def _async_get_device_automations(hass, automation_type, device_id):
     """List device automations."""
     device_registry, entity_registry = await asyncio.gather(
         hass.helpers.device_registry.async_get_registry(),
@@ -79,13 +70,15 @@ async def _async_get_device_automations(hass, fname, device_id):
         config_entry = hass.config_entries.async_get_entry(entry_id)
         domains.add(config_entry.domain)
 
-    entities = async_entries_for_device(entity_registry, device_id)
-    for entity in entities:
-        domains.add(split_entity_id(entity.entity_id)[0])
+    entity_entries = async_entries_for_device(entity_registry, device_id)
+    for entity_entry in entity_entries:
+        domains.add(entity_entry.domain)
 
     device_automations = await asyncio.gather(
         *(
-            _async_get_device_automations_from_domain(hass, domain, fname, device_id)
+            _async_get_device_automations_from_domain(
+                hass, domain, automation_type, device_id
+            )
             for domain in domains
         )
     )
@@ -106,7 +99,7 @@ async def _async_get_device_automations(hass, fname, device_id):
 async def websocket_device_automation_list_actions(hass, connection, msg):
     """Handle request for device actions."""
     device_id = msg["device_id"]
-    actions = await _async_get_device_automations(hass, "async_get_actions", device_id)
+    actions = await _async_get_device_automations(hass, "action", device_id)
     connection.send_result(msg["id"], actions)
 
 
@@ -120,9 +113,7 @@ async def websocket_device_automation_list_actions(hass, connection, msg):
 async def websocket_device_automation_list_conditions(hass, connection, msg):
     """Handle request for device conditions."""
     device_id = msg["device_id"]
-    conditions = await _async_get_device_automations(
-        hass, "async_get_conditions", device_id
-    )
+    conditions = await _async_get_device_automations(hass, "condition", device_id)
     connection.send_result(msg["id"], conditions)
 
 
@@ -136,7 +127,5 @@ async def websocket_device_automation_list_conditions(hass, connection, msg):
 async def websocket_device_automation_list_triggers(hass, connection, msg):
     """Handle request for device triggers."""
     device_id = msg["device_id"]
-    triggers = await _async_get_device_automations(
-        hass, "async_get_triggers", device_id
-    )
+    triggers = await _async_get_device_automations(hass, "trigger", device_id)
     connection.send_result(msg["id"], triggers)
