+++ tests/components/binary_sensor/test_init.py
